Moved per-file scoring orchestration into PerFileScoringTask
* ProcessFile, WriteFeatureDump, LoadSpectra, RunCalibration,
  RunCalibrationScoringPass, SampleLibraryForCalibration, and
  ScoreCalibrationEntry move from AnalysisPipeline into
  PerFileScoringTask as private members (7 methods, ~1500 lines)
* Calibration tuning constants (MIN_SNR_FOR_RT_CAL,
  MIN_COELUTION_CORR_SCORE, CAL_TOP_N_FRAGMENTS, MIN_COELUTION_SPECTRA,
  CAL_FDR_THRESHOLD, ABSOLUTE_MIN_CALIBRATION_POINTS) and the
  CalibrationPassResult nested type travel with them. Shared
  AnalysisPipeline reads CAL_TOP_N_FRAGMENTS via PerFileScoring.PerFileScoringTask.
* The shared scoring engine (RunCoelutionScoring + ScoreWindow +
  ScoreCandidate + feature-compute helpers) stays on AnalysisPipeline
  because Stage6Rescore.cs / PerFileRescoreTask also drives it. Methods
  that PerFileScoringTask now needs to call back into are promoted from
  private to internal: ExtractIsolationWindows, RunCoelutionScoring,
  DeduplicateDoubleCounting, DeduplicatePairs, ExtractTopNFragmentXics,
  CountTop6Matches, plus the static helpers HasTopNFragmentMatch,
  PearsonOverRange, BinarySearchLowerBound, FindNearestMs1
* s_mzmlReadGate promoted to internal (PerFileScoringTask.LoadSpectra
  takes the same gate)
* AnalysisPipeline.cs shrinks by ~1500 lines (now ~3130, down from
  7054 originally -- 56% reduction)
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
